requirements.txt
djangorestframework
django-rest-swagger
django-filter
psycopg2
pylint
coverage
